requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
PySocks==1.7.1
python-telegram-bot==13.6
pytz==2021.1
six==1.16.0
tornado==6.1
tzlocal==2.1
emoji==1.2.0
pymongo==3.11.4
dnspython==2.0.0